+++ .drone/boost-script.sh
@@ -35,3 +35,4 @@ echo '==================================> COMPILE'
 
 cd ../boost-root
 libs/beast/tools/retry.sh libs/beast/tools/build-and-test.sh
+