+++ include/boost/beast/_experimental/test/impl/stream.hpp
@@ -27,53 +27,19 @@ namespace test {
 
 //------------------------------------------------------------------------------
 
-struct stream::service_impl
-{
-    std::mutex m_;
-    std::vector<state*> v_;
-
-    BOOST_BEAST_DECL
-    void
-    remove(state& impl);
-};
-
-class stream::service
-    : public beast::detail::service_base<service>
-{
-    boost::shared_ptr<service_impl> sp_;
-
-    BOOST_BEAST_DECL
-    void
-    shutdown() override;
-
-public:
-    BOOST_BEAST_DECL
-    explicit
-    service(net::execution_context& ctx);
-
-    BOOST_BEAST_DECL
-    static
-    auto
-    make_impl(
-        net::io_context& ctx,
-        test::fail_count* fc) ->
-            boost::shared_ptr<state>;
-};
-
-//------------------------------------------------------------------------------
-
+template<class Executor>
 template<class Handler, class Buffers>
-class stream::read_op : public stream::read_op_base
+class basic_stream<Executor>::read_op : public detail::stream_read_op_base
 {
     using ex1_type =
-        net::io_context::executor_type;
+        executor_type;
     using ex2_type
         = net::associated_executor_t<Handler, ex1_type>;
 
     struct lambda
     {
         Handler h_;
-        boost::weak_ptr<state> wp_;
+        boost::weak_ptr<detail::stream_state> wp_;
         Buffers b_;
 #if defined(BOOST_ASIO_NO_TS_EXECUTORS)
         net::any_io_executor wg2_;
@@ -87,7 +53,7 @@ class stream::read_op : public stream::read_op_base
         template<class Handler_>
         lambda(
             Handler_&& h,
-            boost::shared_ptr<state> const& s,
+            boost::shared_ptr<detail::stream_state> const& s,
             Buffers const& b)
             : h_(std::forward<Handler_>(h))
             , wp_(s)
@@ -95,11 +61,11 @@ class stream::read_op : public stream::read_op_base
 #if defined(BOOST_ASIO_NO_TS_EXECUTORS)
             , wg2_(net::prefer(
                 net::get_associated_executor(
-                  h_, s->ioc.get_executor()),
+                  h_, s->exec),
                 net::execution::outstanding_work.tracked))
 #else // defined(BOOST_ASIO_NO_TS_EXECUTORS)
             , wg2_(net::get_associated_executor(
-                h_, s->ioc.get_executor()))
+                h_, s->exec))
 #endif // defined(BOOST_ASIO_NO_TS_EXECUTORS)
         {
         }
@@ -151,43 +117,44 @@ class stream::read_op : public stream::read_op_base
     };
 
     lambda fn_;
-#if defined(BOOST_ASIO_NO_TS_EXECUTORS)
+#if defined(BOOST_ASIO_USE_TS_EXECUTOR_AS_DEFAULT)
+    net::executor_work_guard<net::any_io_executor> wg1_;
+#else
     net::any_io_executor wg1_;
-#else // defined(BOOST_ASIO_NO_TS_EXECUTORS)
-    net::executor_work_guard<ex1_type> wg1_;
-#endif // defined(BOOST_ASIO_NO_TS_EXECUTORS)
+#endif
 
 public:
     template<class Handler_>
     read_op(
         Handler_&& h,
-        boost::shared_ptr<state> const& s,
+        boost::shared_ptr<detail::stream_state> const& s,
         Buffers const& b)
         : fn_(std::forward<Handler_>(h), s, b)
-#if defined(BOOST_ASIO_NO_TS_EXECUTORS)
-        , wg1_(net::prefer(s->ioc.get_executor(),
+#if defined(BOOST_ASIO_USE_TS_EXECUTOR_AS_DEFAULT)
+        , wg1_(s->exec)
+#else
+        , wg1_(net::prefer(s->exec,
             net::execution::outstanding_work.tracked))
-#else // defined(BOOST_ASIO_NO_TS_EXECUTORS)
-        , wg1_(s->ioc.get_executor())
-#endif // defined(BOOST_ASIO_NO_TS_EXECUTORS)
+#endif
     {
     }
 
     void
     operator()(error_code ec) override
     {
-#if defined(BOOST_ASIO_NO_TS_EXECUTORS)
-        net::post(wg1_, beast::bind_front_handler(std::move(fn_), ec));
-        wg1_ = net::any_io_executor(); // probably unnecessary
-#else // defined(BOOST_ASIO_NO_TS_EXECUTORS)
+#if defined(BOOST_ASIO_USE_TS_EXECUTOR_AS_DEFAULT)
         net::post(wg1_.get_executor(),
             beast::bind_front_handler(std::move(fn_), ec));
         wg1_.reset();
-#endif // defined(BOOST_ASIO_NO_TS_EXECUTORS)
+#else
+        net::post(wg1_, beast::bind_front_handler(std::move(fn_), ec));
+        wg1_ = net::any_io_executor(); // probably unnecessary
+#endif
     }
 };
 
-struct stream::run_read_op
+template<class Executor>
+struct basic_stream<Executor>::run_read_op
 {
     template<
         class ReadHandler,
@@ -195,7 +162,7 @@ struct stream::run_read_op
     void
     operator()(
         ReadHandler&& h,
-        boost::shared_ptr<state> const& in,
+        boost::shared_ptr<detail::stream_state> const& in,
         MutableBufferSequence const& buffers)
     {
         // If you get an error on the following line it means
@@ -209,7 +176,7 @@ struct stream::run_read_op
 
         initiate_read(
             in,
-            std::unique_ptr<read_op_base>{
+            std::unique_ptr<detail::stream_read_op_base>{
             new read_op<
                 typename std::decay<ReadHandler>::type,
                 MutableBufferSequence>(
@@ -220,7 +187,8 @@ struct stream::run_read_op
     }
 };
 
-struct stream::run_write_op
+template<class Executor>
+struct basic_stream<Executor>::run_write_op
 {
     template<
         class WriteHandler,
@@ -228,8 +196,8 @@ struct stream::run_write_op
     void
     operator()(
         WriteHandler&& h,
-        boost::shared_ptr<state> in_,
-        boost::weak_ptr<state> out_,
+        boost::shared_ptr<detail::stream_state> in_,
+        boost::weak_ptr<detail::stream_state> out_,
         ConstBufferSequence const& buffers)
     {
         // If you get an error on the following line it means
@@ -245,7 +213,7 @@ struct stream::run_write_op
         auto const upcall = [&](error_code ec, std::size_t n)
         {
             net::post(
-                in_->ioc.get_executor(),
+                in_->exec,
                 beast::bind_front_handler(std::move(h), ec, n));
         };
 
@@ -281,9 +249,10 @@ struct stream::run_write_op
 
 //------------------------------------------------------------------------------
 
+template<class Executor>
 template<class MutableBufferSequence>
 std::size_t
-stream::
+basic_stream<Executor>::
 read_some(MutableBufferSequence const& buffers)
 {
     static_assert(net::is_mutable_buffer_sequence<
@@ -296,9 +265,10 @@ read_some(MutableBufferSequence const& buffers)
     return n;
 }
 
+template<class Executor>
 template<class MutableBufferSequence>
 std::size_t
-stream::
+basic_stream<Executor>::
 read_some(MutableBufferSequence const& buffers,
     error_code& ec)
 {
@@ -326,7 +296,7 @@ read_some(MutableBufferSequence const& buffers,
         {
             return
                 in_->b.size() > 0 ||
-                in_->code != status::ok;
+                in_->code != detail::stream_status::ok;
         });
 
     // deliver bytes before eof
@@ -340,14 +310,16 @@ read_some(MutableBufferSequence const& buffers,
     }
 
     // deliver error
-    BOOST_ASSERT(in_->code != status::ok);
+    BOOST_ASSERT(in_->code != detail::stream_status::ok);
     ec = net::error::eof;
     return 0;
 }
 
-template<class MutableBufferSequence, BOOST_BEAST_ASYNC_TPARAM2 ReadHandler>
-BOOST_BEAST_ASYNC_RESULT2(ReadHandler)
-stream::
+template<class Executor>
+template<class MutableBufferSequence,
+    BOOST_ASIO_COMPLETION_TOKEN_FOR(void(error_code, std::size_t)) ReadHandler>
+BOOST_ASIO_INITFN_RESULT_TYPE(ReadHandler, void(error_code, std::size_t))
+basic_stream<Executor>::
 async_read_some(
     MutableBufferSequence const& buffers,
     ReadHandler&& handler)
@@ -365,9 +337,10 @@ async_read_some(
             buffers);
 }
 
+template<class Executor>
 template<class ConstBufferSequence>
 std::size_t
-stream::
+basic_stream<Executor>::
 write_some(ConstBufferSequence const& buffers)
 {
     static_assert(net::is_const_buffer_sequence<
@@ -381,9 +354,10 @@ write_some(ConstBufferSequence const& buffers)
     return bytes_transferred;
 }
 
+template<class Executor>
 template<class ConstBufferSequence>
 std::size_t
-stream::
+basic_stream<Executor>::
 write_some(
     ConstBufferSequence const& buffers, error_code& ec)
 {
@@ -425,9 +399,11 @@ write_some(
     return n;
 }
 
-template<class ConstBufferSequence, BOOST_BEAST_ASYNC_TPARAM2 WriteHandler>
-BOOST_BEAST_ASYNC_RESULT2(WriteHandler)
-stream::
+template<class Executor>
+template<class ConstBufferSequence,
+    BOOST_ASIO_COMPLETION_TOKEN_FOR(void(error_code, std::size_t)) WriteHandler>
+BOOST_ASIO_INITFN_RESULT_TYPE(WriteHandler, void(error_code, std::size_t))
+basic_stream<Executor>::
 async_write_some(
     ConstBufferSequence const& buffers,
     WriteHandler&& handler)
@@ -448,11 +424,11 @@ async_write_some(
 
 //------------------------------------------------------------------------------
 
-template<class TeardownHandler>
+template<class Executor, class TeardownHandler>
 void
 async_teardown(
     role_type,
-    stream& s,
+    basic_stream<Executor>& s,
     TeardownHandler&& handler)
 {
     error_code ec;
@@ -477,8 +453,8 @@ async_teardown(
 
 //------------------------------------------------------------------------------
 
-template<class Arg1, class... ArgN>
-stream
+template<class Executor, class Arg1, class... ArgN>
+basic_stream<Executor>
 connect(stream& to, Arg1&& arg1, ArgN&&... argn)
 {
     stream from{
@@ -486,6 +462,34 @@ connect(stream& to, Arg1&& arg1, ArgN&&... argn)
         std::forward<ArgN>(argn)...};
     from.connect(to);
     return from;
+}
+
+namespace detail
+{
+template<class To>
+struct extract_executor_op
+{
+    To operator()(net::any_io_executor& ex) const
+    {
+        assert(ex.template target<To>());
+        return *ex.template target<To>();
+    }
+};
+
+template<>
+struct extract_executor_op<net::any_io_executor>
+{
+    net::any_io_executor operator()(net::any_io_executor& ex) const
+    {
+        return ex;
+    }
+};
+}
+
+template<class Executor>
+auto basic_stream<Executor>::get_executor() noexcept -> executor_type
+{
+    return detail::extract_executor_op<Executor>()(in_->exec);
 }
 
 } // test