+++ include/boost/beast/src.hpp
@@ -28,6 +28,7 @@ the program, with the macro BOOST_BEAST_SEPARATE_COMPILATION defined.
 #include <boost/beast/_experimental/test/impl/error.ipp>
 #include <boost/beast/_experimental/test/impl/fail_count.ipp>
 #include <boost/beast/_experimental/test/impl/stream.ipp>
+#include <boost/beast/_experimental/test/detail/stream_state.ipp>
 
 #include <boost/beast/core/detail/base64.ipp>
 #include <boost/beast/core/detail/sha1.ipp>