+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Add executor rebind to test::stream.
 * Remove floating point arithmetic requirement.
 * Add `cxxstd` to json field.